+++ src/styles/landing/benefitsCarousel.css
@@ -1,4 +1,4 @@
-.benefits_Carousel .swiper-slide {
+.benefits_carousel .swiper-slide {
     text-align: center !important;
     padding: 20px 50px !important;
     width: 650px;
