+++ src/styles/landing/carousel.css
